Add the arithmetic operators +, *, / and %
Addition, multiplication, division and modulo, each an infix operator
over two operands of the same numeric type. Add and Multiply join the
Arithmetic base, so a sum or product past the int range fails the same
way a difference does. Division is total in a second way: `/` and `%`
give an Option of the operand type, none for a zero divisor, and `/` has
one more — PHP_INT_MIN / -1, the one int quotient that overflows and the
one input intdiv() throws for. `%` doesn't share it: PHP_INT_MIN % -1
is 0.

The tokenizer reads the four new operators, the parser folds them into
the cascade — `*`, `/`, `%` at a new multiplicative level tighter than
`+` and `-` — and Precedence prints them back at the level each parses
at. Evaluation cases assert their value with its PHP type, so 0 stays
apart from none and an int result from a float.

Split from #63.
