+++ src/Parser/ExpressionParser.php
@@ -21,7 +21,7 @@ use function str_split;
 /**
  * Expressions are parsed as a cascade of precedence levels, from loosest to tightest binding:
  *
- *     expression → or → and → comparison → additive → unary → postfix → primary
+ *     expression → or → and → comparison → additive → multiplicative → unary → postfix → primary
  *
  * Each level consumes only its own operators and delegates to the next tighter level for its operands. Precedence and
  * associativity are therefore expressed by the call graph, and a level never needs to know which operators sit above
@@ -147,17 +147,51 @@ final class ExpressionParser
     }
 
     /**
-     * a:int - b:int - c:int
+     * a:int + b:int - c:int
      * =====================
+     *
+     * The loop is what makes this level left-associative: each operator folds what came before into its left operand.
+     * {@see self::parseComparison()} is otherwise the same shape with an if in place of the loop, which is what makes
+     * the six comparison operators non-associative; {@see \Eventjet\Ausdruck\Precedence::leftSlot()} reads that one
+     * difference back out when an expression is printed.
      */
     private function parseAdditive(): Expression
     {
-        $left = $this->parseUnary();
-        while ($this->nextToken() === Token::Minus) {
+        $left = $this->parseMultiplicative();
+        while (true) {
+            $build = match ($this->nextToken()) {
+                Token::Plus => $left->add(...),
+                Token::Minus => $left->subtract(...),
+                default => null,
+            };
+            if ($build === null) {
+                return $left;
+            }
             $this->tokens->next();
-            $left = $left->subtract($this->parseUnary());
+            $left = $build($this->parseMultiplicative());
         }
-        return $left;
+    }
+
+    /**
+     * a:int * b:int / c:int
+     * =====================
+     */
+    private function parseMultiplicative(): Expression
+    {
+        $left = $this->parseUnary();
+        while (true) {
+            $build = match ($this->nextToken()) {
+                Token::Asterisk => $left->multiply(...),
+                Token::Slash => $left->divide(...),
+                Token::Percent => $left->modulo(...),
+                default => null,
+            };
+            if ($build === null) {
+                return $left;
+            }
+            $this->tokens->next();
+            $left = $build($this->parseUnary());
+        }
     }
 
     /**
