+++ src/Parser/Tokenizer.php
@@ -62,6 +62,10 @@ final class Tokenizer
                 ')' => Token::CloseParen,
                 ':' => Token::Colon,
                 ',' => Token::Comma,
+                '+' => Token::Plus,
+                '*' => Token::Asterisk,
+                '/' => Token::Slash,
+                '%' => Token::Percent,
                 '[' => Token::OpenBracket,
                 ']' => Token::CloseBracket,
                 '{' => Token::OpenBrace,
