+++ src/Precedence.php
@@ -44,13 +44,14 @@ enum Precedence: int
     case And = 2;
     case Comparison = 3;
     case Additive = 4;
-    case Unary = 5;
+    case Multiplicative = 5;
+    case Unary = 6;
     /**
      * Calls, field accesses and the atomic expressions (literals, variables, lists, structs) all share the tightest
      * level: none of them can have an operand stolen, so none is ever parenthesized as an operand. Lambdas look atomic
      * but are not—see {@see self::Lambda}.
      */
-    case Primary = 6;
+    case Primary = 7;
 
     /**
      * Prints $operand as it appears in an operand slot that the parser reads at $slot, wrapping it in parentheses when
@@ -128,7 +129,8 @@ enum Precedence: int
             Token::And => self::And,
             Token::TripleEquals, Token::NotEquals, Token::CloseAngle, Token::OpenAngle,
             Token::GreaterThanEquals, Token::LessThanEquals => self::Comparison,
-            Token::Minus => self::Additive,
+            Token::Plus, Token::Minus => self::Additive,
+            Token::Asterisk, Token::Slash, Token::Percent => self::Multiplicative,
             default => throw new LogicException(sprintf('%s is not a binary operator', $token->value)),
         };
     }
