+++ tests/unit/ExpressionComparisonTest.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 
 namespace Eventjet\Ausdruck\Test\Unit;
 
+use Eventjet\Ausdruck\Add;
 use Eventjet\Ausdruck\And_;
 use Eventjet\Ausdruck\Call;
 use Eventjet\Ausdruck\ComparisonOperator;
+use Eventjet\Ausdruck\Divide;
 use Eventjet\Ausdruck\Expr;
 use Eventjet\Ausdruck\Expression;
 use Eventjet\Ausdruck\FieldAccess;
@@ -14,6 +16,8 @@ use Eventjet\Ausdruck\Get;
 use Eventjet\Ausdruck\Lambda;
 use Eventjet\Ausdruck\ListLiteral;
 use Eventjet\Ausdruck\Literal;
+use Eventjet\Ausdruck\Modulo;
+use Eventjet\Ausdruck\Multiply;
 use Eventjet\Ausdruck\Negative;
 use Eventjet\Ausdruck\Or_;
 use Eventjet\Ausdruck\Parser\Span;
@@ -49,6 +53,22 @@ final class ExpressionComparisonTest extends TestCase
         yield [
             Expr::subtract(Expr::literal(1), Expr::literal(2)),
             Expr::subtract(Expr::literal(1), Expr::literal(2)),
+        ];
+        yield [
+            Expr::add(Expr::literal(1), Expr::literal(2)),
+            Expr::add(Expr::literal(1), Expr::literal(2)),
+        ];
+        yield [
+            Expr::multiply(Expr::literal(1), Expr::literal(2)),
+            Expr::multiply(Expr::literal(1), Expr::literal(2)),
+        ];
+        yield [
+            Expr::divide(Expr::literal(1), Expr::literal(2)),
+            Expr::divide(Expr::literal(1), Expr::literal(2)),
+        ];
+        yield [
+            Expr::modulo(Expr::literal(1), Expr::literal(2)),
+            Expr::modulo(Expr::literal(1), Expr::literal(2)),
         ];
         yield [
             Expr::gt(Expr::literal(1), Expr::literal(2)),
@@ -209,6 +229,51 @@ final class ExpressionComparisonTest extends TestCase
         yield Subtract::class . ': different type' => [
             Expr::subtract(Expr::literal(1), Expr::literal(2)),
             Expr::literal(1),
+        ];
+        yield Add::class . ': augend is different' => [
+            Expr::add(Expr::literal(1), Expr::literal(2)),
+            Expr::add(Expr::literal(2), Expr::literal(2)),
+        ];
+        yield Add::class . ': addend is different' => [
+            Expr::add(Expr::literal(1), Expr::literal(2)),
+            Expr::add(Expr::literal(1), Expr::literal(1)),
+        ];
+        // The operands match, so only the operator tells the two apart.
+        yield Add::class . ' and ' . Subtract::class => [
+            Expr::add(Expr::literal(1), Expr::literal(2)),
+            Expr::subtract(Expr::literal(1), Expr::literal(2)),
+        ];
+        yield Multiply::class . ': multiplicand is different' => [
+            Expr::multiply(Expr::literal(1), Expr::literal(2)),
+            Expr::multiply(Expr::literal(2), Expr::literal(2)),
+        ];
+        yield Multiply::class . ': multiplier is different' => [
+            Expr::multiply(Expr::literal(1), Expr::literal(2)),
+            Expr::multiply(Expr::literal(1), Expr::literal(1)),
+        ];
+        yield Multiply::class . ' and ' . Divide::class => [
+            Expr::multiply(Expr::literal(1), Expr::literal(2)),
+            Expr::divide(Expr::literal(1), Expr::literal(2)),
+        ];
+        yield Divide::class . ': dividend is different' => [
+            Expr::divide(Expr::literal(1), Expr::literal(2)),
+            Expr::divide(Expr::literal(2), Expr::literal(2)),
+        ];
+        yield Divide::class . ': divisor is different' => [
+            Expr::divide(Expr::literal(1), Expr::literal(2)),
+            Expr::divide(Expr::literal(1), Expr::literal(1)),
+        ];
+        yield Divide::class . ' and ' . Modulo::class => [
+            Expr::divide(Expr::literal(1), Expr::literal(2)),
+            Expr::modulo(Expr::literal(1), Expr::literal(2)),
+        ];
+        yield Modulo::class . ': dividend is different' => [
+            Expr::modulo(Expr::literal(1), Expr::literal(2)),
+            Expr::modulo(Expr::literal(2), Expr::literal(2)),
+        ];
+        yield Modulo::class . ': divisor is different' => [
+            Expr::modulo(Expr::literal(1), Expr::literal(2)),
+            Expr::modulo(Expr::literal(1), Expr::literal(1)),
         ];
         yield Call::class . ': target is different' => [
             Expr::literal(1)->call('foo', Type::int(), []),
