+++ tests/unit/cases/add/in-list-item.txt
@@ -1,0 +1,7 @@
+[a:int + 1, 10]
+-- Input --
+{
+  a: 3,
+}
+-- Output --
+[4, 10]
