+++ tests/unit/cases/divide/by-nonzero-is-some.txt
@@ -1,0 +1,8 @@
+(a:int / b:int).isSome:bool()
+-- Input --
+{
+  a: 1,
+  b: 2,
+}
+-- Output --
+true
