+++ tests/unit/cases/modulo/by-zero-is-none.txt
@@ -1,0 +1,8 @@
+(a:int % b:int).isSome:bool()
+-- Input --
+{
+  a: 5,
+  b: 0,
+}
+-- Output --
+false
