+++ tests/unit/cases/modulo/every-nth.txt
@@ -1,0 +1,7 @@
+(i:int % 3).unwrap:int() === 0
+-- Input --
+{
+  i: 9,
+}
+-- Output --
+true
