+++ tests/unit/cases/modulo/sign-of-dividend.txt
@@ -1,0 +1,8 @@
+(a:int % b:int).unwrap:int()
+-- Input --
+{
+  a: -7,
+  b: 3,
+}
+-- Output --
+-1
