+++ tests/unit/cases/precedence/additive-left-associative.txt
@@ -1,0 +1,9 @@
+a:int - b:int + c:int
+-- Input --
+{
+  a: 10,
+  b: 3,
+  c: 4,
+}
+-- Output --
+11
